# back-end/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Finall project",
    "description": "Back-end Apis",
    "license": {
      "name": "Miu",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": ["http", "https"],
...
  "paths": {
...
    "/admin/farmers": {
      "patch": {
        "tags": ["superuser"],
        "parameters": [
          {
            "name": "",
            "in": "body",
            "description": "admin can activate/deactivate farmer",
            "schema": {
              "$ref": "#/definitions/admin/farmes"
            }
          }
        ],
        "produce": ["application/json"],
        "response": {
          "200": {
            "description": "requested completed sucessfully",
            "schema": {
              "$ref": "#/definitions/admin"
            }
          }
        }
      }
    },
    "/admin/customers": {
      "patch": {
        "tags": ["superuser"],
        "parameters": [
          {
            "name": "",
            "in": "body",
            "description": "admin can  activate/deactivate users",
            "schema": {
              "$ref": "#/definitions/admin/farmes"
            }
          }
        ],
        "produce": ["application/json"],
        "response": {
          "200": {
            "description": "requested completed sucessfully",
            "schema": {
              "$ref": "#/definitions/admin"
            }
          }
        }
      }
    },
...
    "/customers": {
      "patch": {
        "tags": ["customers"],
        "parameters": [
          {
            "name": "rate",
            "in": "body",
            "description": "rate of the customer ",
            "schema": {
              "$ref": "#/definitions/customers"
            }
          }
        ],
        "summary": "customers  ableto rate the farmers  ",
        "response": {
          "200": {
            "description": "ok",
            "schema": {
              "$ref": "#/definitions/customers"
            }
          }
        }
      }
    }
  },
...
}